Fixed #117
We always handle two types of error, exec errors and program stream errors, in the first case we show a notification with the error related information, in the second case we check the exit code and text looking for warnings, if found we will show a warning notification with the warning related information, else an error notification. A new configuration option can suppress the warning notifications to avoid notification spam to the user. This will solve CI errors.